Bugfix: Ensure provider fields are required and allow env variables as an alternative
While allowing users to provide secrets as environment variables instead in
the provider block and ensure that the fields are required and not optional.

Ensure that the fields are required in configureProvider function rather than
the schema definition of the provider since terraform is not aware of the
environment variables.

Update documentation to specify the provider fields are required.