same resources names in tests cause errors
same resource names in different tests case errors since they run
in parallel and if a resource is created in two different tests
and delete by one of the tests, the other test can seem to find
the resource in post-refresh step.

this is a possibility of many more type of errors.